+++ app/app.ts
@@ -1,12 +1,12 @@
+import '@warp-drive/ember/install';
 import Application from '@ember/application';
 import compatModules from '@embroider/virtual/compat-modules';
 import Resolver from 'ember-resolver';
 import loadInitializers from 'ember-load-initializers';
-import config from './config/environment';
+import config from 'winds-mobi-client-web/config/environment';
 import './styles/app.css';
 import '@glint/ember-tsc/types';
 import { setBuildURLConfig } from '@warp-drive/utilities/json-api';
-import '@warp-drive/ember/install';
 
 setBuildURLConfig({
   host: 'https://winds.mobi/api',